--- app/src/main/java/woowacourse/payments/MainActivity.kt
@@ -1,50 +1,0 @@
-package woowacourse.payments
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import woowacourse.payments.ui.addcard.AddCardScreen
-import woowacourse.payments.ui.addcard.AddCardTopbar
-import woowacourse.payments.ui.theme.AndroidpaymentsTheme
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        enableEdgeToEdge()
-        setContent {
-            AndroidpaymentsTheme {
-                Scaffold(
-                    topBar = {
-                        AddCardTopbar()
-                    }
-                ) { padding ->
-                    AddCardScreen(padding)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun Greeting(name: String, modifier: Modifier = Modifier) {
-    Text(
-        text = "Hello $name!",
-        modifier = modifier
-    )
-}
-
-@Preview(showBackground = true)
-@Composable
-fun GreetingPreview() {
-    AndroidpaymentsTheme {
-        Greeting("Android")
-    }
-}
