+++ app/src/main/java/woowacourse/payments/ui/addcard/AddCardScreen.kt
@@ -7,34 +7,22 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.colorResource
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.input.PasswordVisualTransformation
-import androidx.compose.ui.text.input.VisualTransformation
-import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import woowacourse.payments.R
-import woowacourse.payments.model.CardInfo
-import woowacourse.payments.ui.addcard.util.CardNumberTransformation
-import woowacourse.payments.ui.addcard.util.ExpirationDateTransformation
-import woowacourse.payments.ui.addcard.util.PlaceholderTransformation
+import woowacourse.payments.ui.addcard.component.AddCardTopbar
 import woowacourse.payments.ui.component.Card
-import woowacourse.payments.ui.component.CardNumberTextField
-import woowacourse.payments.ui.component.ExpireDateTextField
-import woowacourse.payments.ui.component.OwnerNameTextField
-import woowacourse.payments.ui.component.PasswordTextField
+import woowacourse.payments.ui.addcard.component.CardNumberTextField
+import woowacourse.payments.ui.addcard.component.ExpireDateTextField
+import woowacourse.payments.ui.addcard.component.OwnerNameTextField
+import woowacourse.payments.ui.addcard.component.PasswordTextField
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
 
 @Composable
@@ -49,7 +37,7 @@ fun AddCardScreen(innerPadding: PaddingValues) {
         verticalArrangement = Arrangement.Center,
         horizontalAlignment = Alignment.CenterHorizontally
     ) {
-        Card()
+        Card(CardInfoUiState())
         Column {
             Spacer(modifier = Modifier.height(40.dp))
             CardNumberTextField(
