+++ app/src/main/java/woowacourse/payments/ui/addcard/component/AddCardTopbar.kt
@@ -1,7 +1,5 @@
-package woowacourse.payments.ui.addcard
+package woowacourse.payments.ui.addcard.component
 
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.ArrowBack
 import androidx.compose.material.icons.filled.Check
@@ -11,10 +9,8 @@ import androidx.compose.material3.IconButton
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.dp
 import woowacourse.payments.R
 import woowacourse.payments.ui.theme.AndroidpaymentsTheme
