+++ app/src/main/java/woowacourse/payments/ui/addcard/component/CardNumberTextField.kt
@@ -1,4 +1,4 @@
-package woowacourse.payments.ui.component
+package woowacourse.payments.ui.addcard.component
 
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.OutlinedTextField
@@ -10,32 +10,26 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.input.KeyboardType
 import woowacourse.payments.R
 import woowacourse.payments.ui.addcard.CardInfoUiState
-import woowacourse.payments.ui.addcard.util.ExpirationDateTransformation
+import woowacourse.payments.ui.addcard.util.CardNumberFieldTransformation
 import woowacourse.payments.ui.addcard.util.PlaceholderTransformation
 
 @Composable
-fun ExpireDateTextField(
+fun CardNumberTextField(
     cardInfo: CardInfoUiState,
     modifier: Modifier = Modifier,
     ) {
     OutlinedTextField(
         modifier = modifier,
-        singleLine = true,
-        value = cardInfo.expireDate,
+        value = cardInfo.cardNumber,
         keyboardOptions = KeyboardOptions(keyboardType = KeyboardType.Number),
         onValueChange = {
-            cardInfo.updateCardInfo(expireDate = it)
+            cardInfo.updateCardInfo(cardNumber = it)
         },
-        isError = !cardInfo.isExpirationDateValid,
-        label = { Text(stringResource(R.string.addcard_expire_date_label)) },
-        supportingText = {
-            if (!cardInfo.isExpirationDateValid) {
-                Text(stringResource(R.string.addcard_expire_date_error))
-            }
-        },
-        visualTransformation = if (cardInfo.expireDate.isEmpty()) PlaceholderTransformation(
-            placeholder = stringResource(R.string.addcard_expire_date_placeholder),
+        singleLine = true,
+        label = { Text(stringResource(R.string.addcard_card_number_label)) },
+        visualTransformation = if (cardInfo.cardNumber.isEmpty()) PlaceholderTransformation(
+            placeholder = stringResource(R.string.addcard_card_number_placeholder),
             textColor = colorResource(R.color.payments_placeholder_color)
-        ) else ExpirationDateTransformation(),
+        ) else CardNumberFieldTransformation(),
     )
 }
