+++ app/src/main/res/values/strings.xml
@@ -1,6 +1,8 @@
 <resources>
     <string name="app_name">android-payments</string>
     <string name="payments_topbar_add_card">카드 추가</string>
+    <string name="payments_topbar_all_cards">Payments</string>
+
 
     <!--addcard -->
     <string name="addcard_card_number_label">카드 번호</string>
@@ -12,6 +14,9 @@
     <string name="addcard_owner_name_placeholder">카드에 표시된 이름을 입력하세요</string>
     <string name="addcard_password_label">비밀번호</string>
     <string name="addcard_password_placeholder">0000</string>
+
+    <!--allcards-->
+    <string name="allcards_request_add_card">새로운 카드를 등록해주세요</string>
 
 
     <!--content description-->
